2210: update solution

+++ algorithm_2000-2999/2210-Count Hills and Valleys in an Array/main.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int countHillValley(vector<int>& nums) {
         int cnt = 0;
-        for (size_t i = 1, j = 0; (i + 1) < nums.size(); ++i) {
+        for (int i = 1, j = 0; (i + 1) < nums.size(); ++i) {
             if ((nums[j] > nums[i] && nums[i] < nums[i + 1]) || // valley
                 (nums[j] < nums[i] && nums[i] > nums[i + 1])) { // hill
                 ++cnt;
